pyproject.toml
[project]
name = "leverx-tasks"
version = "0.1.0"
description = "LeverX task related env"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.4.0",
    "python-dotenv>=1.1.1",
]
